# .futurxlab/api-specification.yaml
openapi: 3.0.3
info:
  title: InKnowing API
  description: |
    InKnowing is an AI-powered book dialogue platform that enables users to have conversations with books and book characters.

    ## Business Logic Conservation
    Based on the principle of business logic conservation, this API documentation maintains complete mapping between:
    - User Journey → API Endpoints
    - State Transitions → API Operations
    - Sequence Interactions → API Calls
    - Business Rules → API Responses

    ## Key Features
    - Question-driven and book-driven discovery
    - AI-powered dialogues with books and characters
    - User book upload with vectorization
    - Multi-tier membership system
    - Admin management dashboard

    ## Authentication
    The API uses JWT Bearer tokens for authentication. Tokens are obtained through the login endpoints and must be included in the Authorization header for protected endpoints.

    ## Rate Limiting
    - Free users: 20 requests per day
    - Basic members: 200 requests per month
    - Premium members: 500 requests per month
    - Super members: 1000 requests per month
  version: 1.0.0
  contact:
    name: InKnowing Support
    email: support@inknowing.ai
    url: https://inknowing.ai/support
  license:
    name: Proprietary
    url: https://inknowing.ai/terms
servers:
  - url: https://api.inknowing.ai/v1
    description: Production server
  - url: https://staging-api.inknowing.ai/v1
    description: Staging server
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags:
        - Users
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/membership:
    get:
      tags:
        - Users
      summary: Get membership status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Membership information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/membership/upgrade:
    post:
      tags:
        - Users
      summary: Upgrade membership
      description: |
        Upgrade to a paid membership plan
        Business Logic: Free User → Paid Member State Transition
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipUpgrade'
      responses:
        '200':
          description: Upgrade initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrder'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/quota:
    get:
      tags:
        - Users
      summary: Get dialogue quota
      description: Get remaining dialogue quota for current period
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Quota information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quota'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Search ====================
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        phone:
          type: string
          nullable: true
        wechat_openid:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        nickname:
          type: string
        membership:
          $ref: '#/components/schemas/MembershipType'
        points:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserUpdate:
      type: object
      properties:
        nickname:
          type: string
        avatar:
          type: string
...
    MembershipType:
      type: string
      enum: [free, basic, premium, super]
    Membership:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/MembershipType'
        expires_at:
          type: string
          format: date-time
          nullable: true
        quota_total:
          type: integer
          description: Total quota for current period
        quota_used:
          type: integer
          description: Used quota in current period
        quota_reset_at:
          type: string
          format: date-time
          description: When quota resets
        benefits:
          type: array
          items:
            type: string
    MembershipUpgrade:
      type: object
      required:
        - plan
        - payment_method
      properties:
        plan:
          type: string
          enum: [basic, premium, super]
        duration:
          type: integer
          enum: [1, 3, 6, 12]
          default: 1
          description: Duration in months
        payment_method:
          type: string
          enum: [wechat, alipay]
    Quota:
      type: object
      properties:
        total:
          type: integer
        used:
          type: integer
        remaining:
          type: integer
        reset_at:
          type: string
          format: date-time

    # ==================== Book Schemas ====================
...
    PaymentOrder:
      type: object
      properties:
        order_id:
          type: string
        user_id:
          type: string
        type:
          type: string
          enum: [membership, points]
        amount:
          type: number
        currency:
          type: string
          default: CNY
        status:
          type: string
          enum: [pending, processing, completed, failed, refunded]
        payment_method:
          type: string
          enum: [wechat, alipay]
        payment_url:
          type: string
          description: URL or QR code for payment
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true

    # ==================== Common Schemas ====================
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication

    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication

# API Response Standards
...